consolidate login & redirect code
